Mege ReqPool and ResPool
There is now a ReqResPair which is made up of a request a response and an arena.

Request and responses are not tightly coupled. But you always need both, and
having a single pool removes 50% of the locking.

Also, they can now share a single arena, which again, streamlines things.