Make threadpool responsible for writing
Previously, the worker threads were responsible for reading and writing from
the socket. The threadpool did the routing and executed the user's action.

Now the workers are still responsible for reading, but the threadpool does the
writing.

This has a number benefits:
1 - The thread local buffer can be used for responses. This is, by far, the
fastest allocator, so having it available for responses is a big win.
2 - callback stuff is no longer needed. User code can call res.write() if they
need to data cleaned after the response has been written
3 - I think we can rework some of the allocators. I don't think we need a per
connection arena anymore. It can be thread-local. TODO.

I also reworked res.write, now that it's on the hot path. It uses writev for
the header and body, and it might do fewer system calls in some conditions.