Simplify Response memory usage
Response object fully leverages the recent addition of the thread-local buffer.
The fixed header buffer, fixed body buffer and fallback to the large object pool
(again, for the response only) have been removed in favor of simply using our
"Arena" allocator. This is OK, because our "Arena" is really a fallback of
our thread-local-buffer -> arena.

This greatly simplifies the code and should result in lower minimum and maximum
memory usage, without impacting performance. In fact, with a properly tuned
thread_pool.buffer_size, writing large bodies should be faster.

(The same tricks will not be applicable to the Request object)