Switch to blocking sockets (wha?!)
(This is all in non-blocking mode).

Listening socket remains non-blocking. And we still use kqueue/epoll to be
notified when reads won't block.

But, previously, the code would alternate between blocking and nonblocking.
Earlier, it would switch to blocking mode under specific conditions (like if
the app explictly called write, or for chunked encoding or evens streams.

Then we switched it to try to write and, on WouldBlock, switch to blocking. This
had the huge advantage that applciations now have an explicit response life and
their body can't be freed at the end of the handler.

But since we're using epoll/kqueue to make sure reads don't block, there's no
reasond for this nonblocking<->blocking dance. It's just extra syscalls.