Rework middleware
1 - Global middleware can now be defined. This allows middleware to execute
even when no route is found. This can be useful for things like CORS which
should likely answer an OPTIONS request even on a not found.

2 - When adding middleware to a route, you can now pick between an "append" or
"replace" strategy. This controls how/if the middleware for a route are appended
or replace the middleware for a group and then up the chain to the global.

https://github.com/karlseguin/http.zig/issues/67