Remove ConnManager and merge it all into NonBlocking
Maybe at one point I thought ConnManager would be re-usable between Blocking
and NonBlocking. I don't know. But, that isn't the case.

I really don't want some behavior to be easy to call. Like I don't want
manager.keepalive(conn) to be callable from any place. It has specific thread
safety requirement that need to be thought about, so it shouldn't be easy to
call. If, in the future, we need something like manager.keepalive(conn) in
another place, we need to really think about that. So removing ConnManager and
all of its helpers, and inlining what makes sense, protects sensitive code from
being used in a way that it shouldn't.