On disown, immediately remove socket from loop
When res.disown() is called, the caller is assumed to take over ownership of
the socket. We can't assume anything about the socket after res.disown()
returns, e.g. the caller might immediately close the socket.

Currently, after disown, we transfer the connection back to the worker to remove
it from the loop. This happens via the signal mechanism. So removal happens at
some point in the future. Its possible for the connection to have been closed
by then, which causes a segfault when we try to remove.

When res.disown() is called, we now immediately remove the socket from the loop
before turning ownership of the socket to the caller. This requires some hack
code since the HTTPConn previously had no knowledge/link to the loop.